Rework top-level code to support test profile
This change removes the hard-coded local input paths in the top level workflow
and replaces them with values read from an actual sample sheet. The test
profile has been modified to pass in a minimal sample sheet. The ashlar
module was updated to fix a bug that crashed the versions.yml parser.

The marker csv URL is not ideal but using the one from test-datasets would
have required a more wholesale rewriting of the sample sheet organization.
That rewrite had already been implemented in a separate branch to be merged as
soon as this template update is accepted.