Relax all basicpy mcquant snapshots to summary-only
The inter-system floating point variation in basicpy output has proven to be
too substantial to paper over by just rounding the values. We'll just verify
the output csv files have the right headers and the right number of rows since
that is still stable.

Also fixes a small bug in the summarizeCsv helper where it previously returned
a String instead of the intended Map.