Make markersheet a workflow arg and simplify/correct channels
The markersheet is now a workflow input argument and only parsed once in
PIPELINE_INITIALISATION. The samplesheet channel structure is now standardized
and also normalized regardless of whether input_cycle or input_sample is used.
The main pipeline no longer needs to track which input style is in use.
Channels now make full use of meta maps and groupTuple/join to dramatically
simplify the logic. Some grouping/ordering bugs were fixed as a result.

illumination=manual was removed as an option since this behavior is now
automatically applied when the dfp/ffp columns are filled out in the
samplesheet.

The basicpy and ashlar modules require patches to correct deficiencies in
their original input/output specifications. We will upstream these patches
as soon as possible.