Add test profile for old containers and apptainer+gpu
Old containers with an old version of glibc won't work when running under
apptainer --nv on a newer host OS. This profile uses the more compatible but
experimental --nvccli option instead for modules that are likely to be a
problem. Developers running nf-test locally can select this profile as needed.
Institutional configurations should be aware of the issue as well.